handled error of invalid input length

+++ Test.java
@@ -14,8 +14,10 @@ public class Test {
             System.out.print("$ ");
             String input = scnr.nextLine();
             String[] tokens = input.split(" "); //splitting the input by space
+
             switch(tokens[0]){ //so basically, in every kind of input, the first word is always the command and using a switch statement for better readability
                 case "mkdir": //
+                    if (tokens.length == 1) {System.out.println("You forgot the argument!"); break;}
                     if (tokens[1].equals("-p")) // mkdir -p implementation
                         fs.mkdirp(tokens[2]);
                     else
@@ -24,9 +26,11 @@ public class Test {
                         }
                     break;
                 case "touch":
+                    if (tokens.length == 1) {System.out.println("You forgot the argument!"); break;}
                     fs.touch(tokens[1], Integer.parseInt(tokens[2]));
                     break;
                 case "echo": //first, ill have to get the content, this ill do by locating the two quotation marks and getting the text in between them
+                    if (tokens.length == 1) {System.out.println("You forgot the argument!"); break;}
                     int IndexOfFirstQuot = input.indexOf('"'); // first "
                     int IndexOfSecondQuot = input.lastIndexOf('"'); // second "
                     String content = input.substring(IndexOfFirstQuot + 1, IndexOfSecondQuot); //there is a +1 to exclude writing the quotation mark itself
@@ -37,12 +41,14 @@ public class Test {
                     fs.ls();
                     break;                                    //the rest are just calling the corresponding function according to input
                 case "cd":
+                    if (tokens.length == 1) {System.out.println("You forgot the argument!"); break;}
                     fs.cd(tokens[1]);
                     break;
                 case "pwd":
                     fs.pwd();
                     break;
                 case "rm":
+                    if (tokens.length == 1) {System.out.println("You forgot the argument!"); break;}
                     if (tokens[1].equals("-r")) // rm -r implementation
                         fs.rmr(tokens[2]);
                     else
@@ -52,6 +58,7 @@ public class Test {
                     fs.tree();
                     break;
                 case "grep":
+                    if (tokens.length == 1) {System.out.println("You forgot the argument!"); break;}
                     int IndexOfFirstQuote = input.indexOf('"'); // first "
                     int IndexOfSecondQuote = input.lastIndexOf('"'); // second "
                     String pattern = input.substring(IndexOfFirstQuote + 1, IndexOfSecondQuote); //there is a +1 to exclude writing the quotation mark itself
